refactor(lotto): Lotto 클래스 변경
List 인스턴스 변수를 Set으로 변경
생성자에 로또 유효성검사 추가

+++ src/main/java/lotto/objects/Profit.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class Profit {
     static double getProfit(Map<Object, Long> result, int money) {
-        int sum = 0;
+        Long sum = 0L;
 
         for (Map.Entry<Object, Long> entry : result.entrySet()) {
             sum += getPrize(entry);
